Many changes I introduced on an attempt to remove a Makefile dependency
When I was developing on 2015-07-18, while attending the TCO15 event,
I did a very poor mistake while trying to make it possible to deal with
aruba feature specs. I had to work with chdir-ing and removing
directories, and I think that made something wrong...

I was running and debugging the specs, when I suddenly noticed I cannot
save a file from within vim. Then, I found I have no file in my
repository. I had the directory of dev/clicoder, but the content was
gone. I had nothing to do except being confused and panicked cuz I had
not pushed my changes to GitHub for a long time.

Then, I managed to recover diff from my super useful tool named stdkeep
which stores the output of all commands executed. I remembered executing
`gd master` (git diff master) a little while ago, so I retrieved the
diff from the output of that command, which is the change of this
commit. I'm relieved my changes are not lost, but I'm sad my commit
history is squashed into this one...

My guess is that this was caused when I set the option like
`Aruba.configure { |c| c.working_directory = '.' }` (or
` = expand_path('.')`). I think this is evaluated when in dev/clicoder
directory, though I imagined it would be evaluated when it is executed
(i.e. I guess when file changes occur). It might be that I was not
careful enough, but I think it might be worth thinking to send a pull
request to aruba so that other users don't get this disaster.